add isProposedApiEnabled-util, move checkProposedApiEnabled-util, use utils instead of reading the enableProposedApi-property, https://github.com/microsoft/vscode/issues/129037
