+++ src/vs/workbench/api/common/extHostOutput.ts
@@ -9,8 +9,7 @@ import { URI } from 'vs/base/common/uri';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { IExtHostRpcService } from 'vs/workbench/api/common/extHostRpcService';
-import { IExtensionDescription } from 'vs/platform/extensions/common/extensions';
-import { checkProposedApiEnabled } from 'vs/workbench/services/extensions/common/extensions';
+import { IExtensionDescription, checkProposedApiEnabled } from 'vs/platform/extensions/common/extensions';
 import { ILogger, ILoggerService } from 'vs/platform/log/common/log';
 import { OutputChannelUpdateMode } from 'vs/workbench/contrib/output/common/output';
 import { IExtHostConsumerFileSystem } from 'vs/workbench/api/common/extHostFileSystemConsumer';
