+++ src/vs/workbench/api/common/menusExtensionPoint.ts
@@ -16,6 +16,7 @@ import { DisposableStore } from 'vs/base/common/lifecycle';
 import { ThemeIcon } from 'vs/platform/theme/common/themeService';
 import { Iterable } from 'vs/base/common/iterator';
 import { index } from 'vs/base/common/arrays';
+import { isProposedApiEnabled } from 'vs/platform/extensions/common/extensions';
 
 interface IAPIMenu {
 	readonly key: string;
@@ -633,7 +634,7 @@ commandsExtensionPoint.setHandler(extensions => {
 			title,
 			source: extension.description.displayName ?? extension.description.name,
 			shortTitle,
-			tooltip: extension.description.enableProposedApi ? title : undefined,
+			tooltip: isProposedApiEnabled(extension.description) ? title : undefined,
 			category,
 			precondition: ContextKeyExpr.deserialize(enablement),
 			icon: absoluteIcon
@@ -763,7 +764,7 @@ menusExtensionPoint.setHandler(extensions => {
 				return;
 			}
 
-			if (menu.proposed && !extension.description.enableProposedApi) {
+			if (menu.proposed && !isProposedApiEnabled(extension.description)) {
 				collector.error(localize('proposedAPI.invalid', "{0} is a proposed menu identifier and is only available when running out of dev or with the following command line switch: --enable-proposed-api {1}", entry.key, extension.description.identifier.value));
 				return;
 			}
