+++ src/vs/workbench/services/extensions/common/extensions.ts
@@ -319,12 +319,6 @@ export interface ProfileSession {
 	stop(): Promise<IExtensionHostProfile>;
 }
 
-export function checkProposedApiEnabled(extension: IExtensionDescription): void {
-	if (!extension.enableProposedApi) {
-		throw new Error(`[${extension.identifier.value}]: Proposed API is only available when running out of dev or with the following command line switch: --enable-proposed-api ${extension.identifier.value}`);
-	}
-}
-
 export function toExtension(extensionDescription: IExtensionDescription): IExtension {
 	return {
 		type: extensionDescription.isBuiltin ? ExtensionType.System : ExtensionType.User,
