+++ src/vs/workbench/services/label/common/labelService.ts
@@ -21,6 +21,7 @@ import { match } from 'vs/base/common/glob';
 import { LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { IPathService } from 'vs/workbench/services/path/common/pathService';
+import { isProposedApiEnabled } from 'vs/platform/extensions/common/extensions';
 
 const resourceLabelFormattersExtPoint = ExtensionsRegistry.registerExtensionPoint<ResourceLabelFormatter[]>({
 	extensionPoint: 'resourceLabelFormatters',
@@ -83,7 +84,7 @@ class ResourceLabelFormattersHandler implements IWorkbenchContribution {
 	constructor(@ILabelService labelService: ILabelService) {
 		resourceLabelFormattersExtPoint.setHandler((extensions, delta) => {
 			delta.added.forEach(added => added.value.forEach(formatter => {
-				if (!added.description.enableProposedApi && formatter.formatting.workspaceTooltip) {
+				if (!isProposedApiEnabled(added.description) && formatter.formatting.workspaceTooltip) {
 					// workspaceTooltip is only proposed
 					formatter.formatting.workspaceTooltip = undefined;
 				}
